# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI-NIDS",
    "version": "2.0.0.dev0"
  },
  "paths": {
...
    "/api/settings": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get Runtime Settings",
        "operationId": "get_runtime_settings_api_settings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsOut"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "admin"
        ],
        "summary": "Update Runtime Settings",
        "description": "Change some settings. Unknown keys and out-of-range values are rejected (audit SEC-02).",
        "operationId": "update_runtime_settings_api_settings_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Changes"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/suppressions": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List Rules",
        "operationId": "list_rules_api_suppressions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RuleOut"
                  },
                  "type": "array",
                  "title": "Response List Rules Api Suppressions Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Add Rule",
        "operationId": "add_rule_api_suppressions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/suppressions/{rule_id}": {
      "delete": {
        "tags": [
          "admin"
        ],
        "summary": "Disable Rule",
        "operationId": "disable_rule_api_suppressions__rule_id__delete",
        "parameters": [
          {
            "name": "rule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Rule Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/uploads": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List Uploads",
        "operationId": "list_uploads_api_uploads_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UploadOut"
                  },
                  "type": "array",
                  "title": "Response List Uploads Api Uploads Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Upload Pcap",
        "description": "Upload a pcap/pcapng file for replay. Stored under its own id; the file's type is checked\nby its signature, not its name (audit SEC-04).",
        "operationId": "upload_pcap_api_uploads_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_pcap_api_uploads_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Start Job",
        "operationId": "start_job_api_jobs_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ReplayJob"
                  },
                  {
                    "$ref": "#/components/schemas/TrainJob"
                  },
                  {
                    "$ref": "#/components/schemas/PrepareJob"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List Jobs",
        "operationId": "list_jobs_api_jobs_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobOut"
                  },
                  "title": "Response List Jobs Api Jobs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get Job",
        "operationId": "get_job_api_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}/cancel": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Cancel Job",
        "operationId": "cancel_job_api_jobs__job_id__cancel_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Cancel Job Api Jobs  Job Id  Cancel Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/logs": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Logs",
        "description": "Tail of the API log, or of one job's log (`job:<id>`, e.g. the sensor's).",
        "operationId": "logs_api_logs_get",
        "parameters": [
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(api|job:[0-9a-f]{16})$",
              "default": "api",
              "title": "Source"
            }
          },
          {
            "name": "lines",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 5000,
              "minimum": 1,
              "default": 200,
              "title": "Lines"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogTail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/audit": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Audit Log",
        "operationId": "audit_log_api_audit_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuditEntry"
                  },
                  "title": "Response Audit Log Api Audit Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_pcap_api_uploads_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_pcap_api_uploads_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "params": {
            "additionalProperties": true,
            "type": "object",
            "title": "Params"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "number",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "exit_code": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exit Code"
          },
          "result": {
            "additionalProperties": true,
            "type": "object",
            "title": "Result"
          },
          "created_by": {
            "type": "string",
            "title": "Created By"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "params",
          "status",
          "created_at",
          "started_at",
          "finished_at",
          "exit_code",
          "result",
          "created_by"
        ],
        "title": "JobOut"
      },
...
      "PrepareJob": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "prepare",
            "title": "Kind"
          },
          "dataset": {
            "type": "string",
            "enum": [
              "cicids2017",
              "unsw-nb15"
            ],
            "title": "Dataset"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind",
          "dataset"
        ],
        "title": "PrepareJob"
      },
      "ReplayJob": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "replay",
            "title": "Kind"
          },
          "upload_id": {
            "type": "string",
            "pattern": "^[0-9a-f]{16}$",
            "title": "Upload Id"
          },
          "use_model": {
            "type": "boolean",
            "title": "Use Model",
            "default": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind",
          "upload_id"
        ],
        "title": "ReplayJob"
      },
      "RuleIn": {
        "properties": {
          "type": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 32
              },
              {
                "type": "null"
              }
            ],
            "title": "Type"
          },
          "src": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Src"
          },
          "dst": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Dst"
          },
          "dst_port": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 65535.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Dst Port"
          },
          "until": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Until"
          },
          "reason": {
            "type": "string",
            "maxLength": 500,
            "title": "Reason",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "RuleIn"
      },
      "RuleOut": {
        "properties": {
          "type": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 32
              },
              {
                "type": "null"
              }
            ],
            "title": "Type"
          },
          "src": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Src"
          },
          "dst": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Dst"
          },
          "dst_port": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 65535.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Dst Port"
          },
          "until": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Until"
          },
          "reason": {
            "type": "string",
            "maxLength": 500,
            "title": "Reason",
            "default": ""
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "created_at": {
            "type": "number",
            "title": "Created At"
          },
          "created_by": {
            "type": "string",
            "title": "Created By"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          },
          "alert_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alert Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "created_at",
          "created_by",
          "active",
          "alert_id"
        ],
        "title": "RuleOut"
      },
...
      "SettingsOut": {
        "properties": {
          "values": {
            "$ref": "#/components/schemas/RuntimeSettings"
          },
          "apply": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Apply"
          }
        },
        "type": "object",
        "required": [
          "values",
          "apply"
        ],
        "title": "SettingsOut"
      },
...
      "TrainJob": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "train",
            "title": "Kind"
          },
          "dataset": {
            "type": "string",
            "enum": [
              "cicids2017",
              "unsw-nb15"
            ],
            "title": "Dataset"
          },
          "protocol": {
            "type": "string",
            "enum": [
              "day",
              "random",
              "official"
            ],
            "title": "Protocol",
            "default": "day"
          },
          "features": {
            "type": "string",
            "enum": [
              "full",
              "shared"
            ],
            "title": "Features",
            "default": "full"
          },
          "sample_frac": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Sample Frac"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind",
          "dataset"
        ],
        "title": "TrainJob"
      },
      "UploadOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "sha256": {
            "type": "string",
            "title": "Sha256"
          },
          "format": {
            "type": "string",
            "title": "Format"
          },
          "created_at": {
            "type": "number",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "filename",
          "size",
          "sha256",
          "format",
          "created_at"
        ],
        "title": "UploadOut"
      },
...
    }
  }
}